+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 v2.3.0-nullsafety.0
 
-* Migrate package to null-safety, increase minimum SDK version to 2.12
+* Migrate package to null-safety, increase minimum SDK version to 2.12 (Thanks @simolus3)
 
 v2.2.1 & v2.2.2